% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{cap}
\alias{cap}
\title{World Bank Capacity data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\source{
\href{http://databank.worldbank.org/data/reports.aspx?source=statistical-capacity-indicators}{World Bank Capacity archive}
}
\usage{
data(cap)
}
\description{
Statistical Capacity Indicators provides information on various aspects of 
national statistical systems of developing countries, including an overall 
country-level statistical capacity indicator.
}
\examples{
data(cap)
str(cap)
}
\keyword{datasets}

