% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.submission.R
\name{dc.check.submission}
\alias{dc.check.submission}
\title{Check if a submission to the DataChallenge has the right format.}
\usage{
dc.check.submission(ev)
}
\arguments{
\item{ev}{data.table with specific columns, see below. and predictions in the PREDICTION rows and}
}
\value{
TRUE or FALSE
}
\description{
Your submission should have the following columns 
\itemize{
 \item{"TEAM_ID"}{a unique team ID (character string)}
 \item{"SUBMISSION_ID"}{a unique ID of this subission (character string).}
 \item{"ISO"}{country code of the country for which you are making a submission. You need to make a prediction for 66 countries.}
 \item{"YEAR"}{year for which you are making a submission. This should be either 2015 or 2016 (integer).}
 \item{"PREDICTION"}{Predicted number of epidemic outbreaks in this country and this year. This can be real value.}
}
This function will print a warning if there are any issues with the format of your submission.
}
