% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{dc.evaluate}
\alias{dc.evaluate}
\title{Evaluate the MSE of submitted predictions to the Data Challenge}
\usage{
dc.evaluate(save = FALSE)
}
\arguments{
\item{save}{boolean if to save MSE ranking to file}
}
\value{
data.table with columns F, T, TEAM_ID, SUBMISSION_ID, MSE
}
