% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{devi}
\alias{devi}
\title{World Bank Development Indicators data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\source{
\href{http://databank.worldbank.org/data/reports.aspx?source=world-development-indicators}{World Bank WDI archive}
}
\usage{
data(devi)
}
\description{
World Development Indicators (WDI) is the primary World Bank collection of 
development indicators, compiled from officially recognized international 
sources. It presents the most current and accurate global development data 
available, and includes national, regional and global estimates.
}
\examples{
data(devi)
str(devi)
}
\keyword{datasets}
