% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{emstrain}
\alias{emstrain}
\title{EMS training data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\usage{
data(emstrain)
}
\description{
Number of epidemic disasters per country per year that were reported 
to the WHO between 2010-2016.
}
\examples{
data(emstrain)
str(emstrain)
}
\keyword{datasets}

