% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{World Bank GDP data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\source{
\href{https://data.worldbank.org/indicator/NY.GDP.MKTP.KD?view=map}{World Bank GDP archive}
}
\usage{
data(gdp)
}
\description{
World Bank national accounts data, and OECD National Accounts data files.
}
\examples{
data(gdp)
str(gdp)
}
\keyword{datasets}

