% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{ihr}
\alias{ihr}
\title{IHR data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\source{
\href{http://apps.who.int/gho/data/node.main.IHR?lang=en}{IHR Data Archive}
}
\usage{
data(ihr)
}
\description{
As part of the International Health Regulations monitoring framework,
countries are requested to self assess their capacity in 
responding to epidemic disasters each year. Self-assessments range
from 0 to 100, and are sub-divided into 11 categories.
}
\examples{
data(ihr)
str(ihr)
}
\keyword{datasets}
