% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{nut}
\alias{nut}
\title{World Bank Health and Nutrition data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\source{
\href{http://databank.worldbank.org/data/reports.aspx?source=health-nutrition-and-population-statistics}{World Bank Health and Nutrition archive}
}
\usage{
data(nut)
}
\description{
Key health, nutrition and population statistics gathered from a variety of international sources.
}
\examples{
data(nut)
str(nut)
}
\keyword{datasets}

