% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{pop2013}
\alias{pop2013}
\title{Populations 2013 data set}
\format{An object of class \code{"data.table"}.}
\usage{
data(pop2013)
}
\description{
Population sizes in 2013 data set
}
\examples{
data(pop2013)
str(pop2013)
}
\keyword{datasets}

