% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{pov}
\alias{pov}
\title{World Bank Poverty and Equity data set 2010-2016}
\format{An object of class \code{"data.table"}.}
\source{
\href{http://databank.worldbank.org/data/reports.aspx?source=poverty-and-equity}{World Bank Poverty and Equity archive}
}
\usage{
data(pov)
}
\description{
Poverty and Equity Database.
}
\examples{
data(pov)
str(pov)
}
\keyword{datasets}

