% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{whoreg}
\alias{whoreg}
\title{WHO regions data set}
\format{An object of class \code{"data.table"}.}
\usage{
data(whoreg)
}
\description{
Linking anonymised country identifiers to WHO regions.
}
\examples{
data(whoreg)
str(whoreg)
}
\keyword{datasets}

