% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.for.data.R
\docType{data}
\name{train}
\alias{train}
\title{Epidemic Disasters training data set 2010-2014}
\format{An object of class \code{"data.table"}.}
\usage{
data(train)
}
\description{
Number of epidemic disasters per country per year.
}
\examples{
data(train)
str(train)
}
\keyword{datasets}
